[Update] 実装分離
