+++ include/utils.h
@@ -21,6 +21,10 @@ void imax_dense_format_init(IMAXDenseMatrix *imax_m, int row, int col, int row_p
 void imax_sparse_format_init(IMAXSparseMatrix *imax_sp, int row, int col, int sp_col_blk, int m_col_blk_min);
 void imax_dense_format_init_from_sparse(IMAXDenseMatrix *imax_m, IMAXSparseMatrix *imax_sp, int m_col, int m_col_blk_min);
 void imax_gcn_allocation(IMAXSparseMatrix *imax_sp, IMAXDenseMatrix *imax_h, IMAXDenseMatrix *imax_spmm, IMAXDenseMatrix *imax_w, IMAXDenseMatrix *imax_mm, IMAXDenseMatrix *imax_mm2, IMAXDenseMatrix *imax_mm3);
+void imax_sparse_allocation(IMAXSparseMatrix *imax_sp);
+void imax_sparse_deallocation(IMAXSparseMatrix *imax_sp);
+void imax_dense_allocation(IMAXDenseMatrix *imax_m);
+void imax_dense_deallocation(IMAXDenseMatrix *imax_m);
 
 void convert_imax_dense_format(IMAXDenseMatrix *imax_m, DenseMatrix *m);
 void convert_dense_format(DenseMatrix *m, IMAXDenseMatrix *imax_m);
