+++ src/sparse.c
@@ -1,6 +1,6 @@
 // EMAX6/7 GCN Test Program            //
 // sparse.c                            //
-//         Copyright (C) 2023 by NAIST //
+//         Copyright (C) 2024 by NAIST //
 //          Primary writer: Dohyun Kim //
 //          kim.dohyun.kg7@is.naist.jp //
 #include <stdlib.h>
@@ -35,59 +35,9 @@ void spmm(DenseMatrix *result, SparseMatrix *sp_matrix, DenseMatrix *matrix) {
         }
     }
 }
-
-void mm(DenseMatrix *result, DenseMatrix *a, DenseMatrix *b) {
-    printf("<<CPU>>\n");
-    #ifdef USE_MP
-    #pragma omp parallel for
-    #endif
-    for (int i = 0; i < a->row_size; i++) {
-        #ifdef USE_MP
-        #pragma omp parallel for
-        #endif
-        for (int j = 0; j < b->col_size; j++) {
-            float sum = 0;
-            #ifdef USE_MP
-            #pragma omp parallel for reduction(+:sum)
-            #endif
-            for (int k = 0; k < a->col_size; k++) {
-                sum += a->val[i * a->col_size + k] * b->val[k * b->col_size + j];
-            }
-            result->val[i * b->col_size + j] = sum;
-        }
-    }
-}
-
-void relu(DenseMatrix *result, DenseMatrix *a) {
-    printf("<<CPU>>\n");
-    #ifdef USE_MP
-    #pragma omp parallel for
-    #endif
-    for (int i = 0; i < (a->col_size * a->row_size); i++) {
-        result->val[i] = (a->val[i] > 0) ? a->val[i] : 0;
-    }
-}
-
-void d_relu(DenseMatrix *result, DenseMatrix *a) {
-    printf("<<CPU>>\n");
-    #ifdef USE_MP
-    #pragma omp parallel for
-    #endif
-    for (int i = 0; i < (a->col_size * a->row_size); i++) {
-        result->val[i] = (a->val[i] > 0) ? a->val[i] : 0;
-    }
-}
 #endif
 
-#if !defined(USE_CUDA)
-void allocDenseMatrix(DenseMatrix *matrix) {
-    matrix->val = (float*) malloc(sizeof(float)*matrix->row_size*matrix->col_size);
-}
-
-void freeDenseMatrix(DenseMatrix *matrix) {
-    free(matrix->val);
-}
-
+#ifndef USE_CUDA
 void allocSparseMatrix(SparseMatrix *sp_matrix) {
     sp_matrix->row_p = (int*) malloc(sizeof(int)*(sp_matrix->row_size+1));
     sp_matrix->col_p = (int*) malloc(sizeof(int)*(sp_matrix->nnz));
@@ -109,9 +59,6 @@ void gcn_preprocessing(SparseMatrix *matrix) {
             matrix->val[j] = d_row * d_col;
         }
     }
-    #ifdef USE_CUDA
-        sendSparseMatrixToGPU(matrix);
-    #endif
 }
 
 void spia(SparseMatrix *result, SparseMatrix *sp_matrix) {
@@ -157,82 +104,6 @@ void spia(SparseMatrix *result, SparseMatrix *sp_matrix) {
         if (!is_added) {result->row_p[i+1] = result->row_p[i] + sub + 1;result->col_p[k++]=i;}
         else result->row_p[i+1] = result->row_p[i];
         is_added = 0;
-    }
-}
-
-void softmax(DenseMatrix *result) {
-    for (int i = 0; i < result->row_size; i++) {
-        float max = max_in_array(&(result->val[i * result->col_size]), result->col_size);
-        float log_max = log(max);
-        float sum = 0;
-
-        if (max <= 1) log_max = 0;
-        for (int j = 0; j < result->col_size; j++) {
-            sum += exp(result->val[i * result->col_size + j] + log_max);
-        }
-        for (int j = 0; j < result->col_size; j++) {
-            result->val[i * result->col_size + j] = exp(result->val[i * result->col_size + j] + log_max) / sum;
-        }
-    }
-}
-
-void d_softmax(DenseMatrix *result) {
-    for (int i = 0; i < result->row_size; i++) {
-        for (int j = 0; j < result->col_size; j++) {
-            if (i == j)
-                result->val[i * result->col_size + j] = result->val[i * result->col_size + j] * (1 - result->val[i * result->col_size + j]);
-            else
-                result->val[i * result->col_size + j] = result->val[i * result->col_size + j] * (-result->val[i * result->col_size + j]);
-        }
-    }
-}
-
-float max_in_array(float *array, int size) {
-    int i;
-    float max = -INFINITY;
-
-    for (i = 0; i < size; i++) {
-        if (max < array[i])
-            max = array[i];
-    }
-
-    return max;
-}
-
-void msub(DenseMatrix *result, DenseMatrix *a, DenseMatrix *b) {
-    for (int i = 0; i < a->row_size; i++) {
-        for (int j = 0; j < a->col_size; j++) {
-            result->val[i * a->col_size + j] = a->val[i * a->col_size + j] - b->val[i * a->col_size + j];
-        }
-    }
-}
-
-float mmeans(DenseMatrix *a) {
-    float sum = 0;
-    for (int i = 0; i < a->row_size; i++) {
-        for (int j = 0; j < a->col_size; j++) {
-            sum += a->val[i * a->col_size + j];
-        }
-    }
-    return sum / (a->row_size * a->col_size);
-}
-
-void transpose(DenseMatrix *result, DenseMatrix *a) {
-    for (int i = 0; i < a->row_size; i++) {
-        for (int j = 0; j < a->col_size; j++) {
-            result->val[j * a->row_size + i] = a->val[i * a->col_size + j];
-        }
-    }
-}
-
-void expand_labels(DenseMatrix *labels, Uchar *vlabels)  {
-    for (int i = 0; i < labels->row_size; i++) {
-        for (int j = 0; j < labels->col_size; j++) {
-            if (j == (int)vlabels[i])
-                labels->val[i*labels->col_size + j] = 1.0;
-            else
-                labels->val[i*labels->col_size + j] = 0.0;
-        }
     }
 }
